+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Fix product get_type() incorrectly called
 * Fix WC->cart undefined
 * Fix $el.content for WordPress 5.7
+* New setting for enabling course booking integration
 
 ### 1.5.4: February 23nd, 2021
 * Remove web service urls called in error emails